@@ -21,29 +21,29 @@ dependencies:
     sdk: flutter
   bubble_bottom_bar: ^1.2.0
   circular_check_box: 1.0.1
-  cloud_firestore: ^0.14.1+2
-  connectivity: ^2.0.1
-  cupertino_icons: ^0.1.2
-  equatable: ^1.2.5
-  firebase_auth: ^0.18.1+1
-  firebase_core: ^0.5.0
-  flutter_bloc: ^6.1.0
-  flutter_rounded_date_picker: ^1.0.4
-  flutter_screenutil: ^3.1.1
-  flutter_spinkit: ^4.1.2+1
-  flutter_svg: ^0.19.0
-  google_sign_in: ^4.5.4
-  intl: ^0.16.1
-  material_design_icons_flutter: ^3.3.4595
-  meta: ^1.1.8
-  rxdart: ^0.24.1
-  shared_preferences: ^0.5.12
-  table_calendar: ^2.2.3
+  cloud_firestore: ^1.0.7
+  connectivity: ^3.0.3
+  cupertino_icons: ^1.0.2
+  equatable: ^2.0.0
+  firebase_auth: ^1.1.2
+  firebase_core: ^1.1.0
+  flutter_bloc: ^7.0.0
+  flutter_rounded_date_picker: ^2.0.1
+  flutter_screenutil: ^5.0.0
+  flutter_spinkit: ^5.0.0
+  flutter_svg: ^0.22.0
+  google_sign_in: ^5.0.2
+  intl: ^0.17.0
+  material_design_icons_flutter: ^4.0.5955
+  meta: ^1.3.0
+  rxdart: ^0.26.0
+  shared_preferences: ^2.0.5
+  table_calendar: ^3.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.8.0
+  flutter_launcher_icons: ^0.9.0
 
 flutter_icons:
   android: "launcher_icon"
